+++ R/create_pipeline.R
@@ -4,7 +4,7 @@
 #'
 #' @param pipe_name name of the pipeline and function
 #' @param pipeline_dir directory containing the pipeline scripts
-#' @param frequency how often the pipeline should run (e.g., 1 day, 3 hours, 4 months). Fills in maestroFrequency tag
+#' @param frequency how often the pipeline should run (e.g., 1 day, daily, 3 hours, 4 months). Fills in maestroFrequency tag
 #' @param start_time start time of the pipeline schedule. Fills in maestroStartTime tag
 #' @param tz timezone that pipeline will be scheduled in. Fills in maestroTz tag
 #' @param log_level log level for the pipeline (e.g., INFO, WARN, ERROR). Fills in maestroLogLevel tag