+++ tests/testthat/test-suggest_orch_frequency.R
@@ -73,8 +73,23 @@ test_that("suggest_orch_frequency gives expected errors", {
   )
 
   expect_error(
+    suggest_orch_frequency(data.frame(frequency = "1 day")),
+    regexp = "Schedule is missing required column 'start_time'"
+  )
+
+  expect_error(
     suggest_orch_frequency(data.frame(frequency = c("1 potato", "1 month"), start_time = Sys.time() + 0:1)),
     regexp = "invalid time units"
+  )
+
+  expect_error(
+    suggest_orch_frequency(data.frame(frequency = c("1 day", "1 month"), start_time = c("a", "b"))),
+    regexp = "Schedule columns `start_time`"
+  )
+
+  expect_error(
+    suggest_orch_frequency(data.frame(frequency = c("daily", "1 month"), start_time = Sys.time() + 0:1, skip = "hello")),
+    regexp = "Schedule columns `skip`"
   )
 })