+++ tests/testthat/test_pipelines/test_pipeline_daily_single_bad.R
@@ -1,0 +1,11 @@
+#' Simple mtcars print function
+#'
+#' @maestroFrequency minutely
+#' @maestroStartTime 2024-03-01 09:00:00
+#' @maestroTz UTC
+#'
+#'
+#' @export
+get_mtcars <- function() {
+  mtcars
+}