+++ tests/testthat/test_pipelines/test_pipeline_daily_single_good.R
@@ -1,0 +1,14 @@
+#' Simple mtcars print function
+#'
+#' This is a function that runs every hour starting at
+#' 2024-03-01 09:00:00
+#'
+#' @maestroFrequency daily
+#' @maestroStartTime 2024-03-01 09:00:00
+#' @maestroTz UTC
+#'
+#'
+#' @export
+get_mtcars <- function() {
+  mtcars
+}